.replit.backup/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.1",
    "uvicorn[standard]>=0.34.0",
    "flask-wtf>=1.2.2",
    "alembic>=1.15.2",
    "requests>=2.32.3",
]
